requirements.txt
pyserial>=3.5
matplotlib>=3.7.0
numpy>=1.24.0
pyinstaller>=5.13.0

